fix: move analytics route interception from global setup to auto fixture
