+++ e2e/lib/test.ts
@@ -1,7 +1,9 @@
 /* eslint-disable react-hooks/rules-of-hooks */
 
+import { createUrl } from "@acdh-oeaw/lib";
 import { test as base } from "@playwright/test";
 
+import { env } from "@/config/env.config";
 import { defaultLocale, type Locale } from "@/config/i18n.config";
 import { type AccessibilityScanner, createAccessibilityScanner } from "@/e2e/lib/fixtures/a11y";
 import { createI18n, type I18n, type WithI18n } from "@/e2e/lib/fixtures/i18n";
@@ -9,6 +11,9 @@ import { ImprintPage } from "@/e2e/lib/fixtures/imprint-page";
 import { IndexPage } from "@/e2e/lib/fixtures/index-page";
 
 interface Fixtures {
+	// eslint-disable-next-line @typescript-eslint/no-invalid-void-type
+	beforeEachTest: void;
+
 	createAccessibilityScanner: () => Promise<AccessibilityScanner>;
 	createI18n: (locale: Locale) => Promise<I18n>;
 	createImprintPage: (locale: Locale) => Promise<WithI18n<{ imprintPage: ImprintPage }>>;
@@ -16,6 +21,24 @@ interface Fixtures {
 }
 
 export const test = base.extend<Fixtures>({
+	/** @see https://playwright.dev/docs/test-fixtures#adding-global-beforeeachaftereach-hooks */
+	beforeEachTest: [
+		async ({ context }, use) => {
+			if (env.NEXT_PUBLIC_MATOMO_BASE_URL != null) {
+				const baseUrl = String(
+					createUrl({ baseUrl: env.NEXT_PUBLIC_MATOMO_BASE_URL, pathname: "/matomo.php?**" }),
+				);
+
+				await context.route(baseUrl, (route) => {
+					return route.fulfill({ status: 204, body: "" });
+				});
+			}
+
+			await use();
+		},
+		{ auto: true },
+	],
+
 	async createAccessibilityScanner({ page }, use) {
 		await use(() => {
 			return createAccessibilityScanner(page);
